[GLT-4053] extracted interfaces from Case and related classes
