+++ cardea-data/src/main/java/ca/on/oicr/gsi/cardea/data/CaseRelease.java
@@ -1,84 +1,19 @@
 package ca.on.oicr.gsi.cardea.data;
 
 import java.time.LocalDate;
-import java.util.Objects;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 
-@JsonDeserialize(builder = CaseRelease.Builder.class)
-public class CaseRelease {
+@JsonDeserialize(as = CaseReleaseImpl.class)
+public interface CaseRelease {
 
-  private final String deliverable;
-  private final LocalDate qcDate;
-  private final Boolean qcPassed;
-  private final String qcUser;
-  private final String qcNote;
+  String getDeliverable();
 
-  private CaseRelease(Builder builder) {
-    this.deliverable = Objects.requireNonNull(builder.deliverable);
-    this.qcDate = builder.qcDate;
-    this.qcPassed = builder.qcPassed;
-    this.qcUser = builder.qcUser;
-    this.qcNote = builder.qcNote;
-  }
+  LocalDate getQcDate();
 
-  public String getDeliverable() {
-    return deliverable;
-  }
+  Boolean getQcPassed();
 
-  public LocalDate getQcDate() {
-    return qcDate;
-  }
+  String getQcUser();
 
-  public Boolean getQcPassed() {
-    return qcPassed;
-  }
+  String getQcNote();
 
-  public String getQcUser() {
-    return qcUser;
-  }
-
-  public String getQcNote() {
-    return qcNote;
-  }
-
-  @JsonPOJOBuilder(withPrefix = "")
-  public static class Builder {
-
-    private String deliverable;
-    private LocalDate qcDate;
-    private Boolean qcPassed;
-    private String qcUser;
-    private String qcNote;
-
-    public Builder deliverable(String deliverable) {
-      this.deliverable = deliverable;
-      return this;
-    }
-
-    public Builder qcDate(LocalDate qcDate) {
-      this.qcDate = qcDate;
-      return this;
-    }
-
-    public Builder qcPassed(Boolean qcPassed) {
-      this.qcPassed = qcPassed;
-      return this;
-    }
-
-    public Builder qcUser(String qcUser) {
-      this.qcUser = qcUser;
-      return this;
-    }
-
-    public Builder qcNote(String qcNote) {
-      this.qcNote = qcNote;
-      return this;
-    }
-
-    public CaseRelease build() {
-      return new CaseRelease(this);
-    }
-
-  }
 }
